Revert "Maintenance task: Moved PolynomialFeatures to _polynomial.py"
This reverts commit f76df548ecd31dbe6093d4e8329711197c830542.
